Ooops... Typo.
git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@2520 c046a42c-6fe2-441c-8c8c-71466251a162

+++ fpu/softfloat.c
@@ -1164,7 +1164,7 @@ float32 int64_to_float32( int64 a STATUS_PARAM )
 
 }
 
-float64 uint64_to_float32( uint64 a STATUS_PARAM )
+float32 uint64_to_float32( uint64 a STATUS_PARAM )
 {
     int8 shiftCount;
